requirements.txt
# Core dependencies
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0

# CLIP model
git+https://github.com/openai/CLIP.git

# Image processing
opencv-python>=4.6.0
pillow>=9.0.0

# Configuration management
hydra-core>=1.3.0
omegaconf>=2.3.0
pyyaml>=6.0

# Translation (optional)
google-generativeai>=0.3.0

# GUI (optional)
# tkinter comes with Python

# Testing tools
pytest>=7.0.0
pytest-cov>=3.0.0

# Development tools
black>=22.0.0
flake8>=4.0.0
ipython>=8.0.0

